requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Deprecated==1.2.14
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.23.6
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
requests==2.31.0
setuptools==68.2.2
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.41.2
wrapt==1.16.0
langchain
langchain_openai
langchain_community
fastapi[standard]

